pyproject.toml
[project]
name = "changedetection-prithivi2"
version = "0.1.0"
description = "Wildfire damage detection using Prithivi-EO-2.0 model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "terratorch>=0.1.0",
    "rasterio>=1.3.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "transformers>=4.30.0",
    "timm>=0.9.0",
    "einops>=0.7.0",
    "scipy>=1.11.0",
    "datasets>=4.5.0",
    "huggingface-hub>=0.36.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
